+++ src/clj/withings_client/users.clj
@@ -15,10 +15,6 @@
   [params]
   (db/update-user! params))
 
-(defn update-cid!
-  [params]
-  (db/update-cid-by-name! params))
-
 (defn delete-user!
   [id]
   (db/delete-user! {:id id}))
@@ -27,3 +23,12 @@
   []
   (db/get-users))
 
+;; 2022-08-24 はここから
+(defn user-by-name
+  [name]
+  (db/user-by-name {:name name}))
+
+;; no. update-user-by-name ではないと使い道がない。
+(defn update-cid!
+  [params]
+  (db/update-cid-by-name! params))